+++ lib/BlockchainDB/Lmdb/db_lmdb.h
@@ -34,7 +34,6 @@
 #include "BlockchainDB/BlobDataType.h"
 #include "BlockchainDB/Structures.h"
 #include <boost/thread/tss.hpp>
-#include <boost/variant.hpp>
 
 #define ENABLE_AUTO_RESIZE
 
@@ -189,8 +188,6 @@ public:
 
   virtual CryptoNote::blobdata get_block_blob_from_height(const uint64_t& height) const;
 
-//  virtual TransactionTypes::OutputType getTransactionOutputType(boost::variant<CryptoNote::KeyOutput, CryptoNote::MultisignatureOutput> const & target) const;
-
   virtual uint64_t get_block_timestamp(const uint64_t& height) const;
 
   virtual uint64_t get_top_block_timestamp() const;
@@ -230,16 +227,16 @@ public:
 
   virtual uint64_t get_num_outputs(const uint64_t& amount) const;
 
-  virtual output_data_t get_output_key(const uint64_t& amount, const uint64_t& index);
-  virtual output_data_t get_output_key(const uint64_t& global_index) const;
+  virtual output_data_t get_output_key(const uint64_t& amount, const uint32_t& index);
+  virtual output_data_t get_output_key(const uint32_t& global_index) const;
   virtual void get_output_key(const uint64_t &amount, const std::vector<uint64_t> &offsets, std::vector<output_data_t> &outputs, bool allow_partial = false);
 
   virtual tx_out_index get_output_tx_and_index_from_global(const uint64_t& index) const;
   virtual void get_output_tx_and_index_from_global(const std::vector<uint64_t> &global_indices,
       std::vector<tx_out_index> &tx_out_indices) const;
 
-  virtual tx_out_index get_output_tx_and_index(const uint64_t& amount, const uint64_t& index) const;
-  virtual void get_output_tx_and_index(const uint64_t& amount, const std::vector<uint64_t> &offsets, std::vector<tx_out_index> &indices) const;
+  virtual tx_out_index get_output_tx_and_index(const uint64_t& amount, const uint32_t& index) const;
+  virtual void get_output_tx_and_index(const uint64_t& amount, const std::vector<uint32_t> &offsets, std::vector<tx_out_index> &indices) const;
 
   virtual std::vector<uint64_t> get_tx_amount_output_indices(const uint64_t tx_id) const;
